requirements.txt
aiohttp
fastapi
uvicorn
python-dotenv